+++ skeleton/static/lib.js
@@ -1,0 +1,4 @@
+// Accessing the property via string literal prevents renaming by javascript minifiers which can cause FFI errors
+window['skeleton_lib'] = {
+  log: txt => console.log('Received "' + txt + '" from FFI'),
+};